Local Storage initial version nearly complete
After logging in with a user a reload of the page will read from the local storage first. It will then chain to Characters and Stats. Stopping when the store returns null and checking the server next. If no user is found it will navigate to LoginUserPage. There are some bugs to work out and they will be added to Issues if they are not fixed in the next commit