Fixed Bug introduced in commit 8247b7f
was incorrectly using selector getStatLatestCharId in StatEffects.UPDATE. This was causing it to pass the last stat in the 'array' as the one that was updated instead of whichever one was supposed to be updated. Added selector getStatMetaCharId as a way to get the necessary meta data to update localStorage and not pass unneccesary info. getStatLastestCharId was renamed to getStatAddedCharId for clarity.